Responsi UAS

+++ Customer.cs
@@ -6,6 +6,6 @@ namespace ProjectCustomer
         // PERINTAH: lengkapi property class customer, sesuai petunjuk soal
         public string Kode { get; set; }
         public string Nama { get; set; }
-        public int Utang { get; set; }
+        public double Utang { get; set; }
     }
 }
